Create school landing page with automatic redirection from /school
