+++ CHANGELOG.md
@@ -1,0 +1,29 @@
+# CHANGELOG
+
+## [Новый лендинг школы] - 2024-08-11
+
+### ✨ Новые возможности
+- **🏫 Отдельный лендинг школы** - создана новая структура `/school-landing/`
+- **🔧 Правильные лайауты** - использован собственный `+layout.svelte` без наследования
+- **🔐 Рабочая кнопка админа** - AdminLoginModal правильно интегрирован
+- **📱 Автоматическая переадресация** - `/school` перенаправляет на `/school-landing`
+
+### 🛠️ Структура
+```
+src/routes/school-landing/
+├── +layout.svelte     ← Собственный навбар с кнопкой админа
+├── +page.svelte       ← Hero-секция + превью модулей  
+└── +page.ts           ← Загрузчик данных школы
+```
+
+### 🎯 Преимущества
+- ✅ Полная изоляция от основного лайаута с сайдбаром
+- ✅ Рабочая кнопка "Войти как администратор" 
+- ✅ Загрузка изображений школы из logoUrl
+- ✅ Собственный responsive дизайн
+- ✅ Debug логи для диагностики
+
+### 🧪 Тестирование
+- `http://localhost:5173/school-landing` - новый лендинг
+- `http://localhost:5173/school` - переадресация
+- Кнопка админа с console.log 'School Landing: openAdminModal'
